Fix OCR image text extraction by updating IMAGE_DELIVERY to use container name
The OCR tests were failing because the delivery service couldn't be reached from
the apiv1 container. The issue was that IMAGE_DELIVERY was set to
http://delivery.localhost which resolves to 127.0.0.1 inside containers,
not to the actual delivery service container.

Changed IMAGE_DELIVERY from http://delivery.localhost to http://freegle-delivery
for the apiv1 container so that container-to-container communication works
properly through Docker's internal DNS.